changed uint32_t to uint64_t where it's used for producing combinations, restricted input wordlen to 63

+++ src/comb.h
@@ -11,8 +11,8 @@ class Comb {
   private:
      std::vector<std::pair<uint32_t, char>> m_Id_Lttrs;
   public:
-     Comb(uint32_t val, const Strategy & s) {
-        uint32_t mask = 1;
+     Comb(uint64_t val, const Strategy & s) {
+        uint64_t mask = 1;
         for (uint32_t bid : s.bids()) {
             char letter = (val & mask) + '0';
             m_Id_Lttrs.push_back(std::make_pair(bid, letter));
